old-tests/submission/Infraware/Forms/contents/Forms/Not_Use/select_size.html

<!DOCTYPE HTML>
<html>
 <head>
  <title>Forms</title>
  <script src="../../resources/testharness.js"></script>
  <script src="../../resources/testharnessreport.js"></script>
 </head>
 <body>
	<p>
		<h3>select_size</h3>
	</p>
	
	<li>In select control only 5 items is showed when the document was loaded.</li>
	
	<hr>

	<div id="log"></div>

	<form method="post"
		  enctype="application/x-www-form-urlencoded"
		  action=""
		  name="input_form">
		<p>
		 <label for="unittype">Select unit type:</label>
		 <select id="unittype" name="unittype" multiple>
		  <option value="1"> 1 </option>
		  <option value="2"> 2 </option>
		  <option value="3"> 3 </option>
		  <option value="4"> 4 </option>
		  <option value="5"> 5 </option>
		  <option value="5"> 6 </option>
		  <option value="6"> 7 </option>
		  <option value="6"> 8 </option>
		  <option value="8"> 9 </option>
		  <option value="8"> 10 </option>
		 </select>
		</p>	
	</form>

	<script>
		
		var select = document.getElementById("unittype");
		select.size = 5;

		if (typeof(select.size) == "number") {
			if (select.size != 5) {
				test(function() {
					assert_true(false, "size attribute is not correct.");
				});
			}
		} else {
			test(function() {
				assert_unreached("size attribute is not exist.");
			});
		}

	</script>

 </body>
</html>
